// src/app/module/car-loan/table-sale-car-job/client/dto/table-sale-car-job.dto.ts
export interface SaleCarJobDTO {
    job_number: string;
    job_status: string;
    job_substatus: string;
    bankname: string;
    loan_amount: number;
    loan_type: string;
    numberOfinstallment: number;
    job_type: string;
    loan_detail: string;
    job_close_from: string;
    middle_price: number;
    job_owner_name: string;
    previous_bankname: string;
    liabilities_amount: number;
    previous_contract_years: string;
    previous_contract_remaining: number;
    assigned_from: string;
    monthly_installment: number;
    note: string;
    purposeOfloan: string;
    car_type: string;
    car_year: string;
    car_brand: string;
    car_class: string;
    car_model: string;
    car_plate_number: string;
    car_plate_province: string;
    customer_title: string;
    customer_name: string;
    customer_phone: string;
    convenient_time: string;
    customer_job: string;
    customer_income_amount: string;
    customer_income_file: string;
    customer_province: string;
    customer_bureau: any;
    customer_bureau_detail: any;
    customer_guarantor: any;
    customer_overdue: any;
    create_date: string;
    create_by: string;
    customer_type: any;
    customer_age_job: string;
    customer_have_overdue: string;
    job_create_from: any;
    update_by: string;
    update_date: string;
    job_active: number;
    create_from_team: string;
    user_id: string;
    result_job: string;
    customer_age: string;
    customer_district: string;
    customer_tax_id: any;
    car_number: any;
    close_amount: number;
    customer_surname: string;
    customer_address: string;
    customer_moo: string;
    customer_road: string;
    customer_soi: string;
    customer_area: string;
    customer_zipcode: string;
    car_color: any;
    car_VIN: any;
    customer_no: any;
    customer_email: string;
    customer_birthdate: any;
    customer_ethnicity: any;
    customer_nationality: any;
    transferbook: any;
    car_owner_type: string;
    doc_remark: any;
    interest_rate: number;
    maximum_loan_percent: number;
    have_burden: string;
    assign_date: string;
    car_gear: string;
    car_engine: string;
    previous_monthly_installment: number;
    type: string;
    log_date: string;
    process: string;
    update_name: string;
    job_owner_id: string;
    lasted_status: string;
}

export interface TableSaleCarJobDTO {
    completedJobs: SaleCarJobDTO[];
    inprocessJobs: SaleCarJobDTO[];
    waitJobs: SaleCarJobDTO[];
}